requirements.txt
flask
scikit-learn
pandas
numpy
requests
python-dotenv
joblib

pip install -r requirements.txt